move taintedObjects into separate file
... and remove method receivers where it was not
needed.
