+++ v2/pkg/engine/resolve/loader.go
@@ -180,23 +180,21 @@ type Loader struct {
 
 	validateRequiredExternalFields bool
 
-	// taintedEntities tracks entities fetched with errors.
-	// Later fetches should ignore those entities.
-	taintedEntities map[*astjson.Value]struct{}
+	taintedObjs taintedObjects
 }
 
 func (l *Loader) Free() {
 	l.info = nil
 	l.ctx = nil
 	l.resolvable = nil
-	l.taintedEntities = nil
+	l.taintedObjs = nil
 }
 
 func (l *Loader) LoadGraphQLResponseData(ctx *Context, response *GraphQLResponse, resolvable *Resolvable) (err error) {
 	l.resolvable = resolvable
 	l.ctx = ctx
 	l.info = response.Info
-	l.taintedEntities = make(map[*astjson.Value]struct{})
+	l.taintedObjs = make(taintedObjects)
 	return l.resolveFetchNode(response.Fetches)
 }
 
@@ -367,67 +365,18 @@ func (l *Loader) resolveSingle(item *FetchItem) error {
 func (l *Loader) selectItemsForPath(path []FetchItemPathElement) []*astjson.Value {
 	items := []*astjson.Value{l.resolvable.data}
 	if len(path) == 0 {
-		return l.removeTainted(items)
+		return l.taintedObjs.clearTainted(items)
 	}
 	for i := range path {
 		if len(items) == 0 {
 			break
 		}
-		items = l.selectItems(items, path[i])
+		items = selectItems(items, path[i])
 	}
-	return l.removeTainted(items)
+	return l.taintedObjs.clearTainted(items)
 }
 
-// removeTainted removes taintedEntities from the given items list.
-func (l *Loader) removeTainted(items []*astjson.Value) []*astjson.Value {
-	if len(items) == 0 || len(l.taintedEntities) == 0 {
-		return items
-	}
-	filtered := make([]*astjson.Value, 0, len(items))
-	for _, item := range items {
-		if l.isTaintedEntity(item, 0) {
-			continue
-		}
-		filtered = append(filtered, item)
-	}
-	return filtered
-}
-
-const maximumDepthOfTaintedTraversal = 100
-
-// isTaintedEntity checks if the given `item` is considered isTaintedEntity in the Loader context.
-// Not only the item is being considered, but also its elements if the item is an array,
-// or its values if the item is an object.
-func (l *Loader) isTaintedEntity(item *astjson.Value, depth int) bool {
-	_, ok := l.taintedEntities[item]
-	if ok {
-		return true
-	}
-	if depth > maximumDepthOfTaintedTraversal {
-		// hard limit to prevent stack overflow
-		return false
-	}
-	switch item.Type() {
-	case astjson.TypeArray:
-		for _, elem := range item.GetArray() {
-			if l.isTaintedEntity(elem, depth+1) {
-				return true
-			}
-		}
-	case astjson.TypeObject:
-		obj := item.GetObject()
-		found := false
-		obj.Visit(func(key []byte, value *astjson.Value) {
-			if !found && l.isTaintedEntity(value, depth+1) {
-				found = true
-			}
-		})
-		return found
-	}
-	return false
-}
-
-func (l *Loader) isItemAllowedByTypename(obj *astjson.Value, typeNames []string) bool {
+func isItemAllowedByTypename(obj *astjson.Value, typeNames []string) bool {
 	if len(typeNames) == 0 {
 		return true
 	}
@@ -443,7 +392,7 @@ func (l *Loader) isItemAllowedByTypename(obj *astjson.Value, typeNames []string)
 	return slices.Contains(typeNames, __typeNameStr)
 }
 
-func (l *Loader) selectItems(items []*astjson.Value, element FetchItemPathElement) []*astjson.Value {
+func selectItems(items []*astjson.Value, element FetchItemPathElement) []*astjson.Value {
 	if len(items) == 0 {
 		return nil
 	}
@@ -452,7 +401,7 @@ func (l *Loader) selectItems(items []*astjson.Value, element FetchItemPathElemen
 	}
 
 	if len(items) == 1 {
-		if !l.isItemAllowedByTypename(items[0], element.TypeNames) {
+		if !isItemAllowedByTypename(items[0], element.TypeNames) {
 			return nil
 		}
 
@@ -467,7 +416,7 @@ func (l *Loader) selectItems(items []*astjson.Value, element FetchItemPathElemen
 	}
 	selected := make([]*astjson.Value, 0, len(items))
 	for _, item := range items {
-		if !l.isItemAllowedByTypename(item, element.TypeNames) {
+		if !isItemAllowedByTypename(item, element.TypeNames) {
 			continue
 		}
 		field := item.Get(element.Path...)
@@ -483,7 +432,7 @@ func (l *Loader) selectItems(items []*astjson.Value, element FetchItemPathElemen
 	return selected
 }
 
-func (l *Loader) itemsData(items []*astjson.Value) *astjson.Value {
+func itemsData(items []*astjson.Value) *astjson.Value {
 	if len(items) == 0 {
 		return astjson.NullValue
 	}
@@ -694,7 +643,7 @@ func (l *Loader) mergeResult(fetchItem *FetchItem, res *result, items []*astjson
 			})
 		}
 		if slices.Contains(taintedIndices, 0) {
-			l.taintedEntities[items[0]] = struct{}{}
+			l.taintedObjs.add(items[0])
 		}
 		return nil
 	}
@@ -723,7 +672,7 @@ func (l *Loader) mergeResult(fetchItem *FetchItem, res *result, items []*astjson
 					})
 				}
 				if slices.Contains(taintedIndices, idx) {
-					l.taintedEntities[items[i]] = struct{}{}
+					l.taintedObjs.add(items[i])
 				}
 			}
 		}
@@ -744,7 +693,7 @@ func (l *Loader) mergeResult(fetchItem *FetchItem, res *result, items []*astjson
 			})
 		}
 		if slices.Contains(taintedIndices, i) {
-			l.taintedEntities[items[i]] = struct{}{}
+			l.taintedObjs.add(items[i])
 		}
 	}
 	return nil
@@ -795,120 +744,6 @@ func (l *Loader) appendSubgraphError(res *result, fetchItem *FetchItem, value *a
 	l.ctx.appendSubgraphErrors(res.err, subgraphError)
 
 	return nil
-}
-
-// getTaintedIndices identifies indices of malformed entities based on error paths
-// in the response. It uses errors to find entities that have null value for nullable fields that
-// are required for other fetches. If an error was used to find such an entity, then this error is
-// removed from the errors.
-//
-// The high-level flow of how it is used:
-//
-// 1. Subgraph returns errors for specific entities in the "_entities" array;
-// 2. getTaintedIndices examines error paths like ["_entities", 1, "requiredField"];
-// 3. It validates that the failed field was actually requested for @requires;
-// 4. Marks object at index 1 as "tainted";
-// 5. Later fetches will ignore this tainted object.
-func getTaintedIndices(fetch Fetch, data *astjson.Value, errors *astjson.Value) (indices []int) {
-	info := fetch.FetchInfo()
-	if info == nil {
-		return nil
-	}
-	// build a map to search with
-	requestedForRequires := map[GraphCoordinate]struct{}{}
-	for _, fr := range info.FetchReasons {
-		if fr.IsRequires && fr.Nullable {
-			coord := GraphCoordinate{TypeName: fr.TypeName, FieldName: fr.FieldName}
-			requestedForRequires[coord] = struct{}{}
-		}
-	}
-	if len(requestedForRequires) == 0 {
-		return
-	}
-
-	errorsArray := errors.GetArray()
-	for _, candidate := range errorsArray {
-		errorPath := candidate.Get("path")
-		if astjson.ValueIsNull(errorPath) || errorPath.Type() != astjson.TypeArray {
-			continue
-		}
-		pathItems := errorPath.GetArray()
-		if len(pathItems) == 0 {
-			continue
-		}
-		for i, item := range pathItems {
-			if unsafebytes.BytesToString(item.GetStringBytes()) != "_entities" {
-				continue
-			}
-
-			lastIndex := len(pathItems) - 1
-			// The remaining pathItems should have at least 2 items:
-			if lastIndex-i <= 1 {
-				break
-			}
-			// We have the full path to the failed item.
-			//
-			// For example, if pathItems == ["_entities",0,"nested","a"],
-			// then fieldName would be "a" and [0,"nested","a"] used for selection.
-			fieldName := unsafebytes.BytesToString(pathItems[lastIndex].GetStringBytes())
-			obj, index := selectObjectAndIndex(data, pathItems[i+1:lastIndex])
-			if index == -1 || astjson.ValueIsNull(obj) || obj.Type() != astjson.TypeObject {
-				break
-			}
-
-			// Verify that the value selected by the path is null and extract the enclosing typename.
-			possibleNull := obj.Get(fieldName)
-			if possibleNull == nil || possibleNull.Type() != astjson.TypeNull {
-				break
-			}
-			typeName := unsafebytes.BytesToString(obj.GetStringBytes("__typename"))
-			if typeName == "" {
-				break
-			}
-
-			coord := GraphCoordinate{TypeName: typeName, FieldName: fieldName}
-			if _, ok := requestedForRequires[coord]; !ok {
-				break
-			}
-			indices = append(indices, index)
-			break
-		}
-	}
-	return
-}
-
-// selectObjectAndIndex returns an object and its index using the path as selectors on the response.
-// Path should contain at least the index as the first element. Other elements would lead
-// to deeply nested objects.
-func selectObjectAndIndex(response *astjson.Value, path []*astjson.Value) (*astjson.Value, int) {
-	index := -1
-	if len(path) == 0 {
-		return nil, index
-	}
-	for _, el := range path {
-		var key string
-		switch el.Type() {
-		case astjson.TypeNumber:
-			parsed := el.GetInt()
-			if parsed < 0 {
-				return nil, index
-			}
-			if index == -1 {
-				// index is assigned only once
-				index = parsed
-			}
-			key = strconv.Itoa(parsed)
-		case astjson.TypeString:
-			key = unsafebytes.BytesToString(el.GetStringBytes())
-		default:
-			return nil, -1
-		}
-		response = response.Get(key)
-		if response == nil {
-			return nil, -1
-		}
-	}
-	return response, index
 }
 
 func (l *Loader) mergeErrors(res *result, fetchItem *FetchItem, value *astjson.Value) error {
@@ -1445,7 +1280,7 @@ func (l *Loader) loadSingleFetch(ctx context.Context, fetch *SingleFetch, fetchI
 	res.init(fetch.PostProcessing, fetch.Info)
 	buf := &bytes.Buffer{}
 
-	inputData := l.itemsData(items)
+	inputData := itemsData(items)
 	if l.ctx.TracingOptions.Enable {
 		fetch.Trace = &DataSourceLoadTrace{}
 		if !l.ctx.TracingOptions.ExcludeRawInputData && inputData != nil {
@@ -1511,7 +1346,7 @@ func (l *Loader) loadEntityFetch(ctx context.Context, fetchItem *FetchItem, fetc
 	res.init(fetch.PostProcessing, fetch.Info)
 	buf := acquireEntityFetchBuffer()
 	defer releaseEntityFetchBuffer(buf)
-	input := l.itemsData(items)
+	input := itemsData(items)
 	if l.ctx.TracingOptions.Enable {
 		fetch.Trace = &DataSourceLoadTrace{}
 		if !l.ctx.TracingOptions.ExcludeRawInputData && input != nil {
@@ -1623,7 +1458,7 @@ func (l *Loader) loadBatchEntityFetch(ctx context.Context, fetchItem *FetchItem,
 	if l.ctx.TracingOptions.Enable {
 		fetch.Trace = &DataSourceLoadTrace{}
 		if !l.ctx.TracingOptions.ExcludeRawInputData && len(items) != 0 {
-			data := l.itemsData(items)
+			data := itemsData(items)
 			if data != nil {
 				fetch.Trace.RawInputData, _ = l.compactJSON(data.MarshalTo(nil))
 			}
